fix the merge student_group and student_batch patch

+++ erpnext/patches/v8_0/merge_student_batch_and_student_group.py
@@ -39,15 +39,20 @@ def execute():
 		doc.save()
 
 	# delete the student batch and child-table
-	frappe.delete_doc("DocType", "Student Batch", force=1)
-	frappe.delete_doc("DocType", "Student Batch Student", force=1)
-	frappe.delete_doc("DocType", "Student Batch Instructor", force=1)
+	if frappe.db.table_exists("Student Batch"):
+		frappe.delete_doc("DocType", "Student Batch", force=1)
+	if frappe.db.table_exists("Student Batch Student"):
+		frappe.delete_doc("DocType", "Student Batch Student", force=1)
+	if frappe.db.table_exists("Student Batch Instructor"):
+		frappe.delete_doc("DocType", "Student Batch Instructor", force=1)
 
 	# delete the student batch creation tool
-	frappe.delete_doc("DocType", "Student Batch Creation Tool", force=1)
+	if frappe.db.table_exists("Student Batch Creation Tool"):
+		frappe.delete_doc("DocType", "Student Batch Creation Tool", force=1)
 
 	# delete the student batch creation tool
-	frappe.delete_doc("DocType", "Attendance Tool Student", force=1)
+	if frappe.db.table_exists("Attendance Tool Student"):
+		frappe.delete_doc("DocType", "Attendance Tool Student", force=1)
 
 	# change the student batch to student group in the student attendance
 	frappe.reload_doctype("Student Attendance")
